+++ package.json
@@ -20,7 +20,7 @@
     "nodemon": "^1.19.1"
   },
   "scripts": {
-    "start": "nodemon --exec babel-node index.js",
+    "start": "NODE_ENV=developmmet nodemon --exec babel-node index.js",
     "test": "mocha"
   },
   "repository": {
